+++ test/sixarm_ruby_migration_helper_extensions_test.rb
@@ -10,24 +10,36 @@ require "sixarm_ruby_active_record_migration_mock"
 require "sixarm_ruby_migration_helper_extensions"
 
 class Testing < Minitest::Test
- 
+
  def test_columns
   CreateFoos.up
   for col in [
+    ## Contact
     :email,
-    :freebase,
+    :phone,
+    :uri,
+    :hostname,
+    ## Size
     :height,
     :length,
     :width,
     :depth,
     :mass,
+    ## Geolocation
     :latitude,
     :longitude,
     :altitude,
-    :length,
-    :longitude,
-    :phone,
-    :uri,
+    :elevation,
+    ## Place
+    :iso_3166_1_alpha_2,
+    :iso_3166_1_alpha_3,
+    :iso_3166_1_numeric,
+    ## Admin
+    :timestamp_string,
+    :uuid_string,
+    :zid_string,
+    ## Vendor
+    :freebase,
   ] do
     assert(ActiveRecordMigrationMock.tables['foos'].has_column?(col),"has_column?(:#{col})")
   end
@@ -45,17 +57,32 @@ class CreateFoos < ActiveRecordMigrationMock
 
   def self.up
     create_table 'foos', :force => true do |t|
+      ## Contact
+      t.name
       t.email
+      t.phone
+      t.uri
+      t.hostname
+      ## Size
       t.height
       t.length
       t.width
       t.depth
       t.mass
+      ## Geolocation
       t.latitude
       t.longitude
       t.altitude
-      t.phone
-      t.uri
+      t.elevation
+      ## Place
+      t.iso_3166_1_alpha_2
+      t.iso_3166_1_alpha_3
+      t.iso_3166_1_numeric
+      ## Admin
+      t.timestamp_string
+      t.uuid_string
+      t.zid_string
+      ## Vendor
       t.freebase
     end
   end