+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+* 2019-03-29 5.0.0 Drop automatic index
+* 2019-03-29 4.1.2 Add tests
+* 2019-03-29 4.1.1 Relax UUID index
+* 2019-03-29 4.1.0 Add elevation, hostname, iso countries, admin strings, etc.
 * 2018-06-22 4.0.0 Update to Ruby 2.5, testing gems, etc.
 * 2018-06-13 3.0.1 Fix syntax warning
 * 2018-03-04 3.0.0 Sunset elevation in favor of altitude